requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
aiogram==3.3.0
aiosqlite==0.19.0
pydantic==2.5.0
python-dotenv==1.0.0
python-dateutil==2.8.2
jdatetime==4.1.0
httpx==0.25.2